test(Base): 示例的热重载与异步记账不再赌先后
`samples/base_config` 的热重载段原先只等「文件里的新值进了配置快照」，随后才去读回调置下的标记。
ConfigManager 那一轮是先提交快照、后置回调标记（runReloadTask 里 doReload 之后才调回调），
所以轮询线程有可能在两者之间读到「值已换、回调还没过」，把两条断言之一判成假红——
判据换成两条一起等：回调已落地且新值已可见，循环才放行，之后的断言按构造成立。

`samples/base_log` 的 DropOldest 对账把 countLines 的返回值接成了 std::size_t：那条函数读不到文件时
如实给 -1，接成无符号就变成一个巨大的正数，「落盘 + 丢弃 == 提交条数」这个守恒式在文件根本没读出来时
反而可能凑巧成立。记账改回有符号，负数留在等式里就是失败。

验证：Windows 侧重新编出两个示例后 base_config 连跑 4 次各 22 步全过，run_samples.py 汇总
base_config PASS 22 / base_log PASS 23、0 失败。这两处都是「去掉运气」的改动，改前的形状造不出
确定性的红（竞态窗口在微秒级），如实记在这里；被去掉的是对调度先后的依赖，不是某次可见的失败。

Refs: fbec07d

+++ samples/base_log/main.cpp
@@ -14,6 +14,7 @@
 #include "Base/Log/Sinks/RollingFileSink.h"
 #include "common/SampleSupport.h"
 
+#include <cstdint>
 #include <filesystem>
 #include <fstream>
 #include <iostream>
@@ -248,9 +249,11 @@ namespace
             // 换成正反两条：落盘 + 丢弃必须正好等于提交条数（计数不撒谎），且留在盘上的必须是
             // 最新那条（DropOldest 的定义是丢最旧的）。不要求「确实丢了」：那取决于 worker 抢不抢得到
             // 时间片，两种结果下这两条都成立，所以它们不会因为调度运气而假红
-            const std::size_t landedLineCount  = countLines(droppingPath);
-            const std::size_t droppedCount     = async.droppedEventCount();
-            const std::string landedText       = readWholeFile(droppingPath);
+            // 记账一律走有符号：countLines 读不到文件时如实给 -1，接成 std::size_t 就变成一个巨大的
+            // 正数，守恒式两边因此可以在「文件根本没读出来」时凑巧相等
+            const std::int64_t landedLineCount = countLines(droppingPath);
+            const std::int64_t droppedCount    = static_cast<std::int64_t>(async.droppedEventCount());
+            const std::string  landedText      = readWholeFile(droppingPath);
             Samples::checklist().check(landedLineCount + droppedCount == 50 &&
                                            landedText.find("第 49 条记录") != std::string::npos,
                                        "AsyncSink(DropOldest) 的丢弃计数与落盘条数对得上账，留下的是最新那条");
